+++ rollup.config.js
@@ -1,45 +1,87 @@
-import nodeResolve from '@rollup/plugin-node-resolve';
-import commonjs from '@rollup/plugin-commonjs';
-import babel from '@rollup/plugin-babel';
-import postcss from 'rollup-plugin-postcss';
-import { sizeSnapshot } from 'rollup-plugin-size-snapshot';
-import typescript from '@rollup/plugin-typescript';
+import nodeResolve from '@rollup/plugin-node-resolve'
+import babel from '@rollup/plugin-babel'
+import esbuild from 'rollup-plugin-esbuild'
+import postcss from 'rollup-plugin-postcss'
+import { sizeSnapshot } from 'rollup-plugin-size-snapshot'
+import typescript from '@rollup/plugin-typescript'
+import jsx from 'acorn-jsx'
+import path from 'path'
 
-import pkg from './package.json';
+const createBabelConfig = require('./babel.config')
+const extensions = ['.js', '.ts', '.tsx']
+const { root } = path.parse(process.cwd())
 
-export default {
-  input: './src/index.js',
-  output: [
-    {
-      file: pkg.main,
-      format: 'cjs',
-      exports: 'named',
+function external(id) {
+  return !id.startsWith('.') && !id.startsWith(root)
+}
+
+function getBabelOptions(targets) {
+  return {
+    ...createBabelConfig({ env: (env) => env === 'build' }, targets),
+    extensions,
+    comments: false,
+    babelHelpers: 'bundled',
+  }
+}
+
+function getEsbuild(target) {
+  return esbuild({
+    minify: false,
+    target,
+    tsconfig: path.resolve('./tsconfig.json'),
+  })
+}
+
+function createDeclarationConfig(input, output) {
+  return {
+    input,
+    output: {
+      dir: output,
     },
-    {
-      file: pkg.module,
-      format: 'esm',
-      exports: 'named',
-    },
-  ],
-  external: [
-    'react',
-    'react-dom',
-    'react-dnd',
-    'prop-types',
-    'react-dnd-html5-backend',
-    '@nosferatu500/react-dnd-scrollzone',
-    'react-virtualized',
-    'lodash.isequal',
-  ],
-  plugins: [
-    nodeResolve(),
-    postcss({ extract: 'style.css' }),
-    commonjs({
-      include: 'node_modules/**',
-    }),
-    babel({
-      exclude: 'node_modules/**',
-    }),
-    sizeSnapshot(),
-  ],
-};
+    external,
+    acornInjectPlugins: [jsx()],
+    plugins: [
+      postcss({ extract: 'style.css' }),
+      typescript({ declaration: true, outDir: output })
+    ],
+  }
+}
+
+function createESMConfig(input, output) {
+  return {
+    input,
+    output: { file: output, format: 'esm' },
+    external,
+    plugins: [
+      nodeResolve({ extensions }),
+      postcss({ extract: 'style.css' }),
+      getEsbuild('node12'),
+      sizeSnapshot(),
+    ],
+  }
+}
+
+function createCommonJSConfig(input, output) {
+  return {
+    input,
+    output: { file: output, format: 'cjs', exports: 'named' },
+    external,
+    acornInjectPlugins: [jsx()],
+    plugins: [
+      nodeResolve({ extensions }),
+      postcss({ extract: 'style.css' }),
+      typescript(),
+      babel(getBabelOptions({ ie: 11 })),
+      sizeSnapshot(),
+    ],
+  }
+}
+
+export default function (args) {
+  return [
+    createDeclarationConfig('src/index.ts', 'dist'),
+    createCommonJSConfig('src/index.ts', 'dist/index.js'),
+    createESMConfig('src/index.ts', 'dist/esm/index.js'),
+  ]
+}
+