+++ src/index.ts
@@ -1,3 +1,5 @@
+// @ts-nocheck
+
 import SortableTree, {
   SortableTreeWithoutDndContext,
 } from './react-sortable-tree'