+++ src/utils/dnd-manager.ts
@@ -1,3 +1,5 @@
+// @ts-nocheck
+
 import { DragSource as dragSource, DropTarget as dropTarget } from 'react-dnd'
 import { findDOMNode } from 'react-dom'
 import { getDepth } from './tree-data-utils'